TST: Fix test inheritence / fixturization
Closes https://github.com/pandas-dev/pandas/issues/23746
